fpa-t-backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
google-cloud-bigquery>=3.19.0
httpx>=0.27.0
pytest>=8.1.0
pytest-asyncio>=0.23.0
